Force Ansi ON if on Windows else set to AUTO.
